Correct off-by-one error
Using only moment() to get the current date also returns the current time, meaning that when comparing the app can show that a basho started on the day before shonichi as moment also compares the time and the difference can be less than a day.
As the start and end dates of the bashos contain the date only, the time is set to 0:00, so when getting the current date we ask for the start of the day. This way the same time stamps are compared.